feat(html): product-list responsive

+++ src/main/java/com/example/filrougefo/service/order/OrderService.java
@@ -54,7 +54,7 @@ public class OrderService implements IntOrderService{
     }
 
     @Override
-    public OrderLine addProductToOrder(int productId, long quantity,Client client) {
+    public OrderLine addProductToOrder(int productId, double quantity,Client client) {
 
         Product p = productService.findById(productId);
